python/packages/awsutils/src/main/awsutils/requirements.txt
#
####### requirements.txt #######
#
###### Requirements without Version Specifiers ######
boto3
pyyaml
#
###### Requirements with Version Specifiers ######
#   See https://www.python.org/dev/peps/pep-0440/#version-specifiers
